Use HIPFLAGS for AMD SMI Makefile
