+++ src/components/amd_smi/tests/amdsmi_energy_monotonic.c
@@ -4,17 +4,18 @@
  *          djwoun@gmail.com
  *
  */
-#include <cstdio>
-#include <cstdlib>
+#include <stdio.h>
+#include <stdlib.h>
 #include <unistd.h>
+
 #include "papi.h"
-#include "test_harness.hpp"
+#include "test_harness.h"
 
 int main(int argc, char **argv) {
     harness_accept_tests_quiet(&argc, argv);
-    auto opts = parse_harness_cli(argc, argv);
+    HarnessOpts opts = parse_harness_cli(argc, argv);
 
-    const char* root = std::getenv("PAPI_AMDSMI_ROOT");
+    const char* root = getenv("PAPI_AMDSMI_ROOT");
     if (!root || !*root) {
         SKIP("PAPI_AMDSMI_ROOT not set");
     }
